renamed FindArticles to ArticleListBrowser
